backend/requirements.txt
# Core web backend
flask
gunicorn
python-dotenv

# Database
sqlalchemy
alembic
psycopg[binary]
psycopg_pool

# Google APIs
google-auth
google-auth-oauthlib
google-api-python-client
gspread

# Excel / Spreadsheets
openpyxl
pandas

# Utils
requests
pydantic
pydantic_settings

# Optional (dev/test tools, remove if you want only prod)
# pytest
# pytest-cov
# ruff
# black
# mypy
# pydantic
# pydantic-settings